+++ adam/features/map_maker/map_maker.py
@@ -1,45 +1,43 @@
-import pkg_resources
-
-from distutils.dir_util import copy_tree
-
 from .use_cases import XMLParser
 from .entities import Body
 
 
 class MapMaker:
+    '''
+    the MapMaker class contains all the methods to create a map on top of a given scene
+
+    Parameters
+    ----------
+    filename : str
+        the path to the .xml file to generate
+
+    Methods
+    -------
+    add_bodies:
+        add the designed bodies to the map
+
+    make:
+        exports the map to an .xml file
+
+    add_to:
+        adds the created map to a given scene
+    '''
+
     def __init__(self, filename: str) -> None:
         self.filename: str = filename
         self.name: str = filename.split('/')[-1].split('.')[0]
         self.bodies: list[str] = []
         self.body_names: list[str] = []
 
-    def create_xml(self) -> None:
-        # Wrap file
-        content: str = XMLParser.create_element('worldbody', None, self.bodies)
-        name: str = self.filename.split('/')[-1].split('.')[0]
-        model_attr: str = XMLParser.create_attribute('model', name)
-        content: str = XMLParser.create_element('mujoco', [model_attr], [content])
-
-        with open(self.filename, 'w') as file:
-            file.write(content)
-
-    @staticmethod
-    def export_scene(destination: str) -> None:
-        filename = pkg_resources.resource_filename('adam', 'assets/')
-        copy_tree(filename, destination)
-
-    def add_to_scene(self, filename: str) -> None:
-        line_to_add: str = f'\n\t<include file="{self.name}.xml"/>\n'
-
-        with open(filename, 'r') as f:
-            lines = f.readlines()
-
-        lines.insert(-2, line_to_add)
-
-        with open(filename, 'w') as f:
-            f.writelines(lines)
-
     def add_bodies(self, bodies: list[Body]) -> None:
+        '''
+        add the designed bodies to the map
+
+        Parameters
+        ----------
+        bodies : list[~.entities.Body]
+            the list of bodies to include to the map
+        '''
         for body in bodies:
             if not body.size:
                 raise ValueError(f'the body with id {body.name} must have a size')
@@ -49,6 +47,41 @@ class MapMaker:
 
             self.body_names.append(body.name)
             self._add_body(body)
+
+    def make(self) -> None:
+        '''
+        exports the map to an .xml file
+
+        Notes
+        -----
+        The name of the file is specified at instancing the class
+        '''
+        content: str = XMLParser.create_element('worldbody', None, self.bodies)
+        name: str = self.filename.split('/')[-1].split('.')[0]
+        model_attr: str = XMLParser.create_attribute('model', name)
+        content: str = XMLParser.create_element('mujoco', [model_attr], [content])
+
+        with open(self.filename, 'w') as file:
+            file.write(content)
+
+    def add_to(self, filename: str) -> None:
+        '''
+        adds the created map to a given scene
+
+        Parameters
+        ----------
+        filename : str
+            the path to the .xml file to include the map
+        '''
+        line_to_add: str = f'\n\t<include file="{self.name}.xml"/>\n'
+
+        with open(filename, 'r') as f:
+            lines = f.readlines()
+
+        lines.insert(-2, line_to_add)
+
+        with open(filename, 'w') as f:
+            f.writelines(lines)
 
     def _add_body(self, body: Body) -> None:
         # Geom