+++ adam/features/simulation/entities/data.py
@@ -6,5 +6,16 @@ from .collision_data import CollisionData
 
 @dataclass
 class Data:
+    '''
+    the Data class is a dataclass that contains all the information of the simulation
+
+    Attributes
+    ----------
+    configuration : ~.entities.ConfigurationData
+        the data referring to the configuration of the robot
+
+    collision : ~.entities.CollisionData
+        the data referring to the collisions of the robot
+    '''
     configuration: ConfigurationData
     collision: CollisionData