+++ adam/features/simulation/use_cases/__init__.py
@@ -1,2 +1,5 @@
 from .collision_detector import CollisionDetector
 from .data_manager import DataManager
+from .control_visualizer import ControlVisualizer
+from .viewer import Viewer
+from .controller import Controller